+++ test/test_basic.py
@@ -86,29 +86,18 @@ def test_create(tmp_path: Path) -> None:
     assert not n.nest.file.is_file()
 
 
-@pytest.mark.parametrize("custom_name_func", [True, False])
-def test_nested(tmp_path: Path, custom_name_func: bool) -> None:
-    tmp_path = Path("test")  # needs to not be "." or there's a soundness issue around deferred
+def test_nested(tmp_path: Path) -> None:
+    # tmp_path cannot be ".", what would that mean when paths are resolved?
 
     class PhotoYearFolder(Folder):
         index: Path = Path("index.md")
 
-    if custom_name_func:
-
-        class Photos(Folder):
-            temp: Folder
-            y2024: PhotoYearFolder
-            # provide concrete path which doesn't have y prefix
-            y2025: PhotoYearFolder = PhotoYearFolder("2025")
-            readme: Path = Path("readme.md")
-    else:
-
-        class Photos(Folder):  # type:ignore[no-redef]
-            temp: Folder
-            y2024: PhotoYearFolder
-            # provide concrete path which doesn't have y prefix
-            y2025: PhotoYearFolder = PhotoYearFolder("2025")
-            readme: Path = Path("readme.md")
+    class Photos(Folder):
+        temp: Folder
+        y2024: PhotoYearFolder
+        # provide concrete path which doesn't have y prefix
+        y2025: PhotoYearFolder = PhotoYearFolder("2025")
+        readme: Path = Path("readme.md")
 
     photos = Photos(tmp_path)
     assert isinstance(photos.readme, Path)
@@ -133,7 +122,7 @@ def test_exotic_attributes_okay(tmp_path: Path) -> None:
         class Nested(Folder):
             attrib = lambda x: print(x)  # noqa:E731, PLW0108
 
-        subfolder: A = A("custom_nameed_subfolder")
+        subfolder: A = A("custom_named_subfolder")
 
         readme: Path = Path("readme.txt")
 
@@ -148,94 +137,8 @@ class AsgsYearDir(Folder):
     sa2: Path = Path("SA2.gpkg")
 
 
-class AsgsLayersByYear(FolderPartition[AsgsYearDir]):
-    pass
 
 
-def test_partitioned_folder(path_not_on_disk: Path) -> None:
-    f = AsgsLayersByYear(path_not_on_disk)
-    y2016_dir_subfolder = f.get_subfolder("2016")
-    assert not isinstance(y2016_dir_subfolder, AsgsYearDir)
-    y2016_dir = f.get_partition("2016")
-    assert isinstance(y2016_dir, AsgsYearDir)
-    assert y2016_dir.sa1 == path_not_on_disk / "2016" / "SA1.gpkg"
-
-    # check/ document IO behaviour
-    assert not f.to_path().exists()
-    assert not y2016_dir.sa1.exists()
-    f.create()
-    assert f.to_path().is_dir()
-    # child under partition can't be materialised
-    assert not y2016_dir.sa1.exists()
-
-
-def test_enumerated_partitioned_folder(path_not_on_disk: Path) -> None:
-    # repeat test with EnumeratedFolderPartition
-
-    class EnumeratedAsgsLayersByYear(EnumeratedFolderPartition[AsgsYearDir]):
-        partition_names = ("2016", "2021")
-
-    f = EnumeratedAsgsLayersByYear(path_not_on_disk)
-    y2016_dir_subfolder = f.get_subfolder("2016")
-    assert not isinstance(y2016_dir_subfolder, AsgsYearDir)
-    y2016_dir = f.get_partition("2016")
-    assert isinstance(y2016_dir, AsgsYearDir)
-    assert y2016_dir.sa1 == path_not_on_disk / "2016" / "SA1.gpkg"
-
-    # check/ document IO behaviour
-    assert not f.to_path().exists()
-    assert not y2016_dir.sa1.exists()
-    f.create()
-    assert f.to_path().is_dir()
-    # listed child under partition can be materialised
-    assert y2016_dir.to_path().is_dir()
-    assert f.get_subfolder("2021").to_path().exists()
-    assert not f.get_subfolder("2023").to_path().exists()
-    with pytest.raises(NameError):
-        f.get_partition("2023")
-
-
-def test_prefixed_enumerated_partitioned_folder(path_not_on_disk: Path) -> None:
-    # repeat test with EnumeratedFolderPartition
-
-    class EnumeratedAsgsLayersByYear(EnumeratedFolderPartition[AsgsYearDir]):
-        partition_prefix = "year="
-        partition_names = ("2016", "2021")
-
-    f = EnumeratedAsgsLayersByYear(path_not_on_disk)
-    y2016_dir_subfolder = f.get_subfolder("year=2016")  # conforms but wrong method
-    assert not isinstance(y2016_dir_subfolder, AsgsYearDir)
-    y2016_dir = f.get_partition("year=2016")  # explicit prefix
-    assert y2016_dir.sa1 == path_not_on_disk / "year=2016" / "SA1.gpkg"
-    y2016_dir2 = f.get_partition("2016")  # implicit prefix
-    assert y2016_dir == y2016_dir2  # attrs equality implies equal
-    assert y2016_dir != y2016_dir_subfolder
-    assert isinstance(y2016_dir2, AsgsYearDir)
-    assert y2016_dir2.sa1 == path_not_on_disk / "year=2016" / "SA1.gpkg"
-
-    # check/ document IO behaviour
-    assert not f.to_path().exists()
-    assert not y2016_dir.sa1.exists()
-    f.create()
-    assert f.to_path().is_dir()
-    # listed child under partition can be materialised
-    assert y2016_dir.to_path().is_dir()
-    assert f.get_subfolder("year=2021").to_path().exists()
-    assert not f.get_subfolder("year=2023").to_path().exists()
-    with pytest.raises(NameError):
-        f.get_partition("year=2023")
-
-
-def test_enumerated_subfolder_logical(path_not_on_disk: Path) -> None:
-    class EnumeratedAsgsLayersByYear(EnumeratedFolderPartition[AsgsYearDir]):
-        partition_prefix = "year="
-        partition_names = ("2016", "2021")
-
-    f = EnumeratedAsgsLayersByYear(path_not_on_disk)
-
-    assert type(f.get_subfolder("foo")) == Folder  # Shouldn't be AsgsYearDir, doesn't conform # noqa: E721
-    assert type(f.get_subfolder("foo", subfolder_class=AsgsYearDir)) == AsgsYearDir  # noqa: E721
-    assert type(f.get_partition("2016")) == AsgsYearDir  # noqa: E721
 
 
 def test_attrs_subclass_post_init(path_not_on_disk: Path) -> None:
